Make Ed25519*Key structs use Vec<u8> instead of String
The Ed25519PublicKey and Ed25519PrivateKey structs store the key
material as a String. That unfortunately does not work as the actual
data may not be valid UTF-8.
With this change we adjust the types to use Vec<u8> instead of String
and, hence, avoiding any encoding related checks. This change resolves
issue #5.