Export SignatureBlob and Passphrase typedefs
The SignatureBlob and Passphrase types, despite contained in the
exported Message enum, are private to the crate. While that is not
strictly wrong, it can make for some non-obvious client code that just
passes around Vec<u8> and String objects.
With this change we make those two typedefs available to user code.