Accept AsRef<Path> in Agent::run_unix signature
The run_unix method on the Agent trait currently expects an str
describing the path of the domain socket to use. That appears to be an
unnecessary restriction. For one, tokio's UnixListener::bind (the only
user of this path) is more flexible, accepting anything satisfying
AsRef<Path>. But perhaps more importantly it forces clients to convert
into a string first, which in turns checks for valid UTF-8 which,
depending on the system, may not actually be an invariant enforced for
paths by the file system.
With this change we adjust the method's signature to accept any
AsRef<Path>.