[bug] Fix Extension encoded length
Extension is odd compared to other messages because its `Unparsed`
payload does not include a length field. The implementation of
`Encode::encode` for `Extension` accounts for this by writing the bytes
directly from the `Vec<u8>` member of `Unparsed`, but the implementation
of `Encode::encoded_len` for `Extension` did not account for the 4 extra
bytes (a `u32`) that `<Vec<u8> as Encode>::encode` normally prepends.

Therefore, an Extension message on the wire claims it has 4 more bytes
than its payload actually contains, which hangs most SSH agent server
implementations waiting for the 4 additional bytes to arrive.