Remaining types in `proto::message` to modules
All types are `pub use` in `mod proto`, so can be accessed at the same
module path as before this change. Also exposed a `pub use
crate::proto::Result` in `proto:message` to match the `type Result`
alias that used to be there in order to avoid that paper cut of a
breaking change.

Also refactor tests into the modules that the test, except for a few
left in `proto::message` that use `Request` and `Response` as entry
points to test the parsing implementation in many submodules.

Signed-off-by: Ross Williams <ross@ross-williams.net>