unparsed: expose content of the extension
Unparsed is intended to hold the value of the extension.
Specification calls for:
```
    byte             SSH_AGENTC_EXTENSION
    string           extension type
    byte[]           extension request-specific contents
```

But does not mandates for content to be serialized according to ssh
regular serialization format (as in implements `ssh_encoding::Decode`).

The `Unparsed::From` implementation allows for content to be formatted
in any serialization, but you can't read data back unless it was encoded
according to SSH serialization.

This partially reverts https://github.com/wiktor-k/ssh-agent-lib/pull/67

Signed-off-by: Arthur Gautier <arthur.gautier@arista.com>